Use 32-bit guest os and share root of ioio git
* Instead of cloning another git within the virtual machine the host
  git is now available on /ioio.
* Use 32 bit to save us from installing 32 bit libs.
* Some more checks in bootstrap.sh in order to allow "vagrant provision"
  to be called.

"The manual":
    * Install Vagrant: https://docs.vagrantup.com/v2/installation/
    * Install VirtualBox: https://www.virtualbox.org/
    * cd ioio/firmware/vagrant
    * cp PATH-TO-LINUX-MPLABC_COMPILER/mplabc30-v3.30c-linux-installer.run .
    * vagrant up
    * git clone https://github.com/ytai/ioio.git
    * cd /ioio
    * tools/make-all all

+++ firmware/vagrant/bootstrap.sh
@@ -12,26 +12,50 @@ apt-get install curl -y
 # We need unzip below
 apt-get install unzip -y
 
+file /usr/bin/file | grep -q 64-bit
 # 32 bit library support is needed for compiler and its installer
-apt-get install ia32-libs -y
+if [ $? -eq 0 ]; then
+  apt-get install ia32-libs -y
+else
+  echo "No need to install 32-bit libraries since we are running under 32 bit os"
+fi
 
-# Install c30 compiler unattended mode. 
-# mplab c30 compiler for Linux does not seem to be available oneline anymore so
-# you need to supply a local copy in the /vagrant folder
-/vagrant/mplabc30-v3.30c-linux-installer.run -- --mode unattended
-cd /opt/microchip/mplabc30/v3.30c/bin/bin
-ln -s ../c30_device.info
-ln -s ../device_files
-cd -
+if [ ! -d "/opt/microchip/mplabc30/v3.30c" ]; then
+  # Install c30 compiler unattended mode.
+  # mplab c30 compiler for Linux does not seem to be available oneline anymore so
+  # you need to supply a local copy in the /vagrant folder
+  if [ ! -f "/vagrant/mplabc30-v3.30c-linux-installer.run" ]; then
+    echo "You must have a local copy of mplabc30-v3.30c-linux-installer.run in your vagrant folder"
+  else
+    /vagrant/mplabc30-v3.30c-linux-installer.run -- --mode unattended
+    cd /opt/microchip/mplabc30/v3.30c/bin/bin
+    ln -s ../c30_device.info
+    ln -s ../device_files
+    cd -
+  fi
 
-# Download and install MPLAB X
-curl -O http://ww1.microchip.com/downloads/en/DeviceDoc/MPLABX-v2.30-linux-installer.zip
-unzip -q MPLABX-v2.30-linux-installer.zip 
-chmod a+x MPLABX-v2.30-linux-installer.sh
-./MPLABX-v2.30-linux-installer.sh -- --mode unattended
-# Set up path correctly
-echo 'PATH=/opt/microchip/mplabx/mplab_ide/bin/:$PATH'  > /etc/profile.d/microchip_mplabx.sh
+else
+  echo "mplabc30 already installed"
+fi
 
-# Install git
-apt-get install git -y
+# Download and install MPLAB X if not there alreay
+if [ ! -d "/opt/microchip/mplabx/mplab_ide/bin" ]; then
+  # Use local copy if available
+  if [ -f "/vagrant/MPLABX-v2.30-linux-installer.zip" ]; then
+    echo "Using local copy of MPLABX-v2.30-linux-installer.zip"
+    cp /vagrant/MPLABX-v2.30-linux-installer.zip .
+  else
+    echo "Downloading MPLABX-v2.30-linux-installer.zip"
+    curl -O http://ww1.microchip.com/downloads/en/DeviceDoc/MPLABX-v2.30-linux-installer.zip
+  fi
+  unzip -q MPLABX-v2.30-linux-installer.zip
+  chmod u+x MPLABX-v2.30-linux-installer.sh
+  ./MPLABX-v2.30-linux-installer.sh -- --mode unattended
+  rm MPLABX-v2.30-linux-installer.zip
+  rm MPLABX-v2.30-linux-installer.sh
+  # Set up path correctly
+  echo 'PATH=/opt/microchip/mplabx/mplab_ide/bin/:$PATH'  > /etc/profile.d/microchip_mplabx.sh
+else
+  echo "MPLAB X already installed"
+fi
